+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug fixes
 
 - Handle Elixir 1.12 bug on single atom ASTs
+- Handle atoms prefixed with `Elixir.` properly
 
 ## v0.1.0 (2021-09-19)